requirements.txt
pandas
numpy
jupyterlab
scikit-learn==1.3.2
joblib
mlflow
django
psycopg2-binary